+++ src/libstd/std.rc
@@ -99,7 +99,6 @@ pub mod base64;
 pub mod rl;
 pub mod workcache;
 pub mod bigint;
-pub mod biguint;
 
 #[cfg(unicode)]
 mod unicode;
